tarnish-server/requirements.txt
amqp==2.2.2
backports-abc==0.5
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto3==1.7.10
botocore==1.10.10
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
EditorConfig==0.12.2
futures==3.2.0
idna==2.6
jmespath==0.9.3
jsbeautifier==1.7.5
kombu==4.1.0
pika==0.11.2
pycurl==7.43.0.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.11.0
tornado==5.0.2
#tornado-celery==0.3.5
tornado-redis==2.4.18
urllib3==1.22
vine==1.1.4
expiringdict==1.1.4